✨ feat(service): add DIMSE-N request handling and runtime safety guards
- Add N-EVENT-REPORT, N-GET, N-SET, N-ACTION, N-CREATE, and N-DELETE service handlers
- Prevent shutdown from blocking forever by timing out request finalization
- Serialize client connection state changes and avoid double closing service-managed connections
- Cap malformed parser sequences and fragment streams to stop unbounded item reads
- Fix signed 16-bit grayscale decoding, preserve presentation context mutations, and bound film box allocation

+++ pkg/imaging/render/export.go
@@ -90,9 +90,10 @@ func (e *ImageExporter) ExportGrayscale(
 			case 1:
 				pixelValue = float64(pixelData[pixelIndex])
 			case 2:
-				// 16-bit pixel
+				// 16-bit pixel — use uint16 for low-byte assembly to avoid
+				// sign-extension on unsigned bytes ≥ 128, then cast to signed.
 				if isSigned {
-					pixelValue = float64(int16(pixelData[pixelIndex]) | int16(pixelData[pixelIndex+1])<<8)
+					pixelValue = float64(int16(uint16(pixelData[pixelIndex]) | uint16(pixelData[pixelIndex+1])<<8))
 				} else {
 					pixelValue = float64(uint16(pixelData[pixelIndex]) | uint16(pixelData[pixelIndex+1])<<8)
 				}
